+++ include/AstraCAN.h
@@ -33,18 +33,7 @@
 
 typedef TwaiCAN AstraCAN;
 
-void printCANframe(CanFrame& frame) {
-    Serial.print("CAN: ");
-    Serial.print(frame.identifier, HEX);
-    Serial.print(" [");
-    Serial.print(frame.data_length_code);
-    Serial.print("] ");
-    for (int i = 0; i < frame.data_length_code; i++) {
-        Serial.print(frame.data[i], HEX);
-        Serial.print(" ");
-    }
-    Serial.println();
-}
+void printCANframe(CanFrame& frame);
 
 
 #    elif defined(CORE_TEENSY)
